It's a fresh start.
You could say that from the code side the game has been completely rebuilt.
Let's start with the units.  They have a common abstract model, both player and enemies. That will give in the future interesting opportunities, such as taking control of the enemy unit. Control became better after the controller was used as a contextual state machine. Now each state has conditions for changing states. In the game there are weapons, it is 2 types of revolvers. They have the same implementation, but the player and the enemy have their own nuances in controlling these weapons. Weapons and can be used through the hand, which only appeared in this implementation. We don't have a lot of animations, so making a whole arm behind the blank allowed you to do interesting things with the object in your hand. Thus the weapon in the hand can rotate towards the scope.
We got a scope so we could get the direction for the projectile at the weapon. Not only that. Through the sight you can now control the weapon in your hand, firing or reloading. Left and right mouse buttons. We now have a unit spavener. It is easy to use it, you need to set up points on the stage, they will be the targets for the spavin unit. The spavner has a model in which you need to specify the prefab of the unit and set a couple of parameters for spawning. Takzhest simple UI window that tracks the weapon the number of rounds in the player.